Moved 8 Functions into langtonant,js and modified main.js to work with langtonant.js
