+++ katas/langtonant/jhon/main.js
@@ -1,4 +1,3 @@
-import {Board} from './board.js';
-let boardMain = new Board(process.argv[2], process.argv[3]);
-boardMain.printBoard();
-boardMain.langtonantFunction(process.argv[4]);
+import {Langtonant} from './langtonant.js';
+
+new Langtonant(process.argv[2], process.argv[3], process.argv[4]);
